+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - UTF8 character support for argument and command input (by Prool http://mud.kharkov.org, proolix@gmail.com)
 - RGB 8-bit color codes are implemented. Check do_who() for "wch->level" coloring. Color names are taken from: https://www.w3schools.com/colors/
 - Immortal characters can see Room, Mob, Object VNUMs when Holylight is active.
+- Autoexit setting removed. From now on exits will always be shown.